requirements.txt
antiorm==1.2.1
Babel==2.14.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
Flask==3.0.1
Flask-Login==0.6.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
pyodbc==5.0.1
Werkzeug==3.0.1
